fixup! docs: add metadata documentation to httpClient
